Changed @Inject to @Redirect
REFACTOR: Changed the @Inject to @Redirect in the ClientPlayNetworkHandler mixin. This is to remove the default particles from the totem of undying
FEAT: Readded said removed default particles by adding it to the mixin file. This is to allow the config to change if the default particles are to be shown or not
Signed-off-by: Superkat32 <superkat1806@gmail.com>